Added onClick Delete Confirmation Functionality

+++ resources/views/edit.blade.php
@@ -49,7 +49,7 @@
         <td>{{ $target->ranking }}</td>
         <td class="action_td">
             <a href="/edit/{{$target->id}}"><button type="button" class="btn btn-info fa fa-edit"> Edit</button></a>
-            <a href="/delete/{{$target->id}}"><button type="button" class="btn btn-danger fa fa-trash">Delete</button></a>
+            <a href="/delete/{{$target->id}}"><button type="button" class="btn btn-danger fa fa-trash" onclick="return confirm('Are You Sure You Want To Delete!');">Delete</button></a>
         </td>
       </tr>
       @endforeach
